pyproject.toml
[project]
name = "graphrag-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.3.0",
    "litellm>=1.67.5",
    "neo4j>=5.28.1",
    "neo4j-graphrag>=1.7.0",
]
